Register with unity-settings-daemon, make it work on Ubuntu 16.04
With this change we let i3-gnome work on default Ubuntu
installations by disguising it as Unity (this, among other things
lets unity-control-center work).

This means that from now on, i3-gnome is meant to be used on a
default Ubuntu install with Unity and we'll probably need to
provide separate configuration files and packages for vanilla-GNOME
users.